Move combineNSigma in a dedicated HF PID header.
